Added difficulty panel, fifth difficulty level
Added a difficulty panel, which allows for selection of more than 3 difficulty levels. Old difficulty panel is now used as a separate combat adjuster.

+++ game/server/globals.cpp
@@ -23,5 +23,5 @@
 #include "tier0/memdbgon.h"
 
 Vector			g_vecAttackDir;
-int				g_iSkillLevel;
+float				g_iSkillLevel;
 bool			g_fGameOver;
